requirements.txt
# Core Framework
streamlit==1.30.0
python-dotenv==1.0.0

# PDF Processing
pdfplumber==0.10.3
PyMuPDF==1.24.14  # Versão com wheels pré-compilados para Windows

# HTTP Client & API
httpx==0.25.2
pydantic==2.12.5
pydantic-settings==2.12.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Security & Authentication
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
cryptography==41.0.7
pyOpenSSL==23.3.0

# Validation & Utilities
validate-docbr==1.10.0
python-multipart==0.0.6

# Progress & UI
tqdm==4.66.1

# Logging
loguru==0.7.2

# Testing (opcional)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
